FileUploaderField component, refs erm-337
