+++ lib/DocumentsFieldArray/DocumentsFieldArray.js
@@ -15,7 +15,7 @@ import {
 } from '@folio/stripes/components';
 
 import EditCard from '../EditCard';
-import FileUploaderField from './FileUploaderField';
+import FileUploaderField from '../FileUploaderField';
 import withKiwtFieldArray from '../withKiwtFieldArray';
 
 class DocumentsFieldArray extends React.Component {
