+++ lib/DocumentsFieldArray/tests/DocumentsFieldArray-test.js
@@ -7,7 +7,6 @@ import { mountWithContext } from '../../../tests/helpers';
 import TestForm from '../../../tests/TestForm';
 import DocumentsFieldArray from '../DocumentsFieldArray';
 import DocumentsFieldArrayInteractor from './interactor';
-import FileUploaderInteractor from '../../FileUploader/tests/interactor';
 
 chai.use(spies);
 const { expect, spy } = chai;
@@ -20,13 +19,11 @@ const onUploadFile = spy(() => {
   const myResponse = new Response(blob, init);
   return Promise.resolve(myResponse);
 });
-const onUploadFileError = spy(() => Promise.resolve({ error: 400, message: 'Upload Error' }));
 
 const addDocBtnLabel = 'Add document';
 const isEmptyMessage = 'No documents';
 
 const interactor = new DocumentsFieldArrayInteractor();
-const uploaderInteractor = new FileUploaderInteractor();
 
 describe('DocumentsFieldArray', () => {
   beforeEach(async () => {
@@ -172,151 +169,6 @@ describe('DocumentsFieldArray', () => {
 
     it('sets the category value', () => {
       expect(interactor.category).to.equal('foo');
-    });
-  });
-
-  describe('attaching a file', () => {
-    beforeEach(async () => {
-      await interactor.clickAddButton();
-    });
-
-    it('shows the title', () => {
-      expect(uploaderInteractor.isTitleVisible).to.be.true;
-    });
-
-    it('shows the button', () => {
-      expect(uploaderInteractor.isButtonVisible).to.be.true;
-    });
-
-    it('shows the children', () => {
-      expect(uploaderInteractor.isChildrenVisible).to.be.true;
-    });
-
-    describe('dropping multiple files', () => {
-      beforeEach(async () => {
-        await uploaderInteractor.dragEnter();
-        await uploaderInteractor.dropMultiple();
-      });
-
-      it('should not call onUploadFile', () => {
-        expect(onUploadFile).to.not.be.called;
-      });
-    });
-
-    describe('entering the drag area', () => {
-      beforeEach(async () => {
-        await uploaderInteractor.dragEnter();
-      });
-
-      it('shows the title', () => {
-        expect(uploaderInteractor.isTitleVisible).to.be.true;
-      });
-
-      it('hides the button', () => {
-        expect(uploaderInteractor.isButtonVisible).to.be.false;
-      });
-
-      it('hides the children', () => {
-        expect(uploaderInteractor.isChildrenVisible).to.be.false;
-      });
-
-      describe('leaving the drag area', () => {
-        beforeEach(async () => {
-          await uploaderInteractor.dragLeave();
-        });
-
-        it('shows the title', () => {
-          expect(uploaderInteractor.isTitleVisible).to.be.true;
-        });
-
-        it('shows the button', () => {
-          expect(uploaderInteractor.isButtonVisible).to.be.true;
-        });
-
-        it('shows the children', () => {
-          expect(uploaderInteractor.isChildrenVisible).to.be.true;
-        });
-      });
-
-      describe('dropping a file', () => {
-        beforeEach(async () => {
-          await uploaderInteractor.drop();
-        });
-
-        it('calls onUploadFile', () => {
-          expect(onUploadFile).to.be.called();
-        });
-
-        it('renders file name', () => {
-          expect(interactor.file.hasFilename).to.be.true;
-          expect(interactor.file.filename).to.equal(uploadedFile.name);
-        });
-
-        it('renders uploaded date', () => {
-          expect(interactor.file.hasUploaded).to.be.true;
-        });
-
-        it('renders delete button', () => {
-          expect(interactor.file.hasDeleteBtn).to.be.true;
-        });
-
-        describe('clicking the filename link', () => {
-          beforeEach(async () => {
-            await interactor.file.download();
-          });
-
-          expect('calls onDownloadFile', () => {
-            expect(onDownloadFile).to.be.called();
-          });
-        });
-
-        describe('deleting an uploaded file', () => {
-          beforeEach(async () => {
-            await interactor.file.delete();
-          });
-
-          it('does not render file name', () => {
-            expect(interactor.file.hasFilename).to.be.false;
-          });
-
-          it('does not render uploaded date', () => {
-            expect(interactor.file.hasUploaded).to.be.false;
-          });
-
-          it('does not render delete button', () => {
-            expect(interactor.file.hasDeleteBtn).to.be.false;
-          });
-        });
-      });
-    });
-  });
-
-  describe('error on upload', () => {
-    beforeEach(async () => {
-      await mountWithContext(
-        <TestForm>
-          <FieldArray
-            addDocBtnLabel={addDocBtnLabel}
-            component={DocumentsFieldArray}
-            isEmptyMessage={isEmptyMessage}
-            name="docs"
-            onDownloadFile={onDownloadFile}
-            onUploadFile={onUploadFileError}
-          />
-        </TestForm>
-      );
-    });
-
-    describe('dropping a file', () => {
-      beforeEach(async () => {
-        await interactor.clickAddButton();
-        await uploaderInteractor.dragEnter();
-        await uploaderInteractor.drop();
-      });
-
-      it('should show an error', () => {
-        expect(uploaderInteractor.hasError).to.be.true;
-      });
     });
   });
