+++ lib/DocumentsFieldArray/tests/interactor.js
@@ -6,23 +6,10 @@ import {
   focusable,
   interactor,
   isPresent,
-  scoped,
   selectable,
   text,
   value,
 } from '@bigtest/interactor';
-
-@interactor class FileUploaderFieldInteractor {
-  hasFilename = isPresent('[data-test-fuf-name]');
-  hasUploaded = isPresent('[data-test-fuf-uploaded]');
-  hasDeleteBtn = isPresent('[data-test-fuf-delete]');
-
-  filename = text('[data-test-fuf-name]');
-  uploaded = text('[data-test-fuf-uploaded]');
-  delete = clickable('[data-test-fuf-delete]');
-
-  download = clickable('[data-test-fuf-name]');
-}
 
 @interactor class DocumentsFieldArrayInteractor {
   static defaultScope = '[data-test-documents-field-array]';
@@ -51,7 +38,6 @@ import {
   blurUrl = blurrable('[data-test-document-field-url]');
 
   category = value('[data-test-document-field-category]');;
-  file = scoped('[data-test-document-field-file]', FileUploaderFieldInteractor)
 
   clickAddButton = clickable('[data-test-documents-field-array-add-button]');
   addDocBtnLabel = text('[data-test-documents-field-array-add-button]');
