+++ lib/FileUploader/FileUploader.js
@@ -34,6 +34,7 @@ export default class FileUploader extends React.Component {
     rejectClassName: PropTypes.string,
     style: PropTypes.object,
     title: PropTypes.node.isRequired,
+    uploadButtonAriaLabel: PropTypes.string,
     uploadButtonText: PropTypes.node.isRequired,
     uploadInProgress: PropTypes.bool,
     uploadInProgressText: PropTypes.node,
