+++ lib/FileUploaderField/FileUploaderFieldView.js
